# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is math solving API.",
        "title": "Mvthbot API",
        "contact": {
            "name": "API Support",
            "url": "http://www.t.me/un000nu",
            "email": "stephan@un000nu.pw"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "un000nu.mvth.pw",
    "basePath": "/api/v1",
    "paths": {
...
        "/solve/{equation}": {
            "post": {
                "description": "Solves equation given in url param",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solve",
                    "api"
                ],
                "summary": "Basic equations solving",
                "parameters": [
                    {
                        "type": "string",
                        "description": "equation for solve (2+2) encoded in LF",
                        "name": "equation",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}